packages/desktop/linux_requirements.txt
bottle==0.12.25
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clr-loader==0.2.6
greenlet==3.0.3
idna==3.6
packaging==23.2
proxy_tools==0.1.0
psutil==5.9.7
pycairo==1.25.1
pycparser==2.21
PyGObject==3.46.0
PySide2==5.13.2
python-dotenv==1.0.0
pythonnet==3.0.3
pywebview==4.4.1
QtPy==2.4.1
requests==2.31.0
shiboken2==5.13.2
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
